Little fix

+++ src/Helper.cpp
@@ -3,7 +3,7 @@
 DWORD Helpers::GetPointerAddress(DWORD baseAddress, std::vector<DWORD> offsets)
 {
 	DWORD addr = baseAddress;
-	for (int i = 0; i < offsets.size(); i++)
+	for (size_t i = 0; i < offsets.size(); i++)
 	{
 		addr = *(DWORD*)addr;
 		addr += offsets[i];
